Object-based wrapper for reporting
I'm trying to add some features to a reporting base class
such that all child reports can inherit the features. Besides
some standardization the main features added so far are
jQuery based sorting & styling and result caching.

The department and product movement reports have the
best commenting.

I realize "classlib2.0" is kind of ridiculous. I'm not entirely
happy with the module system I was tinkering with in "class-lib"
and it will go away eventually. In the long run I'd like the
backend class and plugin system to mimic the lane both for the
sake of consistency and because I think the lane model is
better.