Session state cleanup & standardization
I removed "warned" and "warnBoxType" so that warnings using
boxMsg2.php are uniformly controlled by "msgrepeat". I gave
that page a URL parameter "quiet" to control whther or not
it should beep the scale.

I removed "inactMem" because it was used in only one place
and was redundant with "memType".

I removed "chargetender" because it was redundant with "chargeTotal".

I removed "receiptType" because it wasn't actually doing anything.

I commented every setting that's still in CoreState.php. I think
I've pared it down to only settings that 1) still do something in
the current code and 2) belong in the browser session.